Изменено: *Обичное поле ввода аудитории изменено на поле ввода с автоматическим заполнением. dialog_discipline_options.xml; ScheduleOfDayActivity.setDisciplineOptions(); ScheduleBuilderActivity.java; В xml файле изменен элемент управления. В ScheduleOfDayActivity.setDisciplineOptions() дописано програмная часть для дееспособности автозаполнения. В ScheduleBuilderActivity.java создание дополнительные публичные статичные поля auditoryOfDisciplines, buildingOfDisciplines; *В типе дисциплине добавлено "Лабораторная". strings.xml; Discipline.java; В strings.xml в массиве type_of_discipline добавлен новая ячейка: "Laboratory". Discipline.java добавлено контрольное поле. *В строковых значениях удалено activity_ScheduleOfDay_dialog_types. *Тип дисциплины теперь измеряется как целочисленное значение(int). Discipline.java. Так же было изменены методы: пуе и сет.

+++ app/src/main/java/com/example/schedule/Objects/DayOfWeek.java
@@ -59,7 +59,7 @@ public class DayOfWeek implements Parcelable
     public ArrayList<Discipline> getDisciplines(){
         return disciplines;
     }
-    public byte getCount(){return (byte) disciplines.size();}
+    public int getCount(){return disciplines.size();}
     public Discipline getDiscipline(int position){
         return disciplines.get(position);
     }
